move the src client around and updated codegen
