--- contrib/kafka/source/PROJECT
@@ -1,3 +1,0 @@
-version: "1"
-domain: eventing.knative.dev
-repo: knative.dev/eventing-kafka
